pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ux-journey-scraper"
version = "0.1.0"
description = "Record and analyze user journeys with automated UX guidelines validation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Rishabh Patel", email = "rishabh@example.com"}
]
keywords = ["ux", "user-experience", "journey", "scraper", "playwright", "testing", "guidelines", "accessibility"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "ecommerce-ux-guidelines>=1.2.0",
    "urllib3>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/rishabhpatelgofynd/ux-journey-scraper"
Documentation = "https://github.com/rishabhpatelgofynd/ux-journey-scraper#readme"
Repository = "https://github.com/rishabhpatelgofynd/ux-journey-scraper"
"Bug Tracker" = "https://github.com/rishabhpatelgofynd/ux-journey-scraper/issues"

[project.scripts]
ux-journey = "ux_journey_scraper.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["ux_journey_scraper*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
